discount points for marking

+++ educouchsystem/src/main/java/com/educouch/educouchsystem/model/Quiz.java
@@ -44,7 +44,7 @@ public class Quiz extends Assessment implements Serializable {
         this.quizQuestions = new ArrayList<>();
     }
 
-    public Quiz(String title, String description, Double maxScore, Double discountPointForAssessment,Integer discountPointToTopPercent, Date startDate, Date endDate, Boolean hasTimeLimit, Integer timeLimit, Boolean isAutoRelease) {
+    public Quiz(String title, String description, Double maxScore, Long discountPointForAssessment,Integer discountPointToTopPercent, Date startDate, Date endDate, Boolean hasTimeLimit, Integer timeLimit, Boolean isAutoRelease) {
         super(title, description, maxScore, startDate, endDate,discountPointForAssessment,discountPointToTopPercent);
         this.hasTimeLimit = hasTimeLimit;
         this.isAutoRelease = isAutoRelease;
